update feature to current

+++ src/features/flights/flights.repo.ts
@@ -4,24 +4,37 @@ export async function get_flight(trip_id: number) {
   const sql = `
     SELECT 
       f.flight_id, 
-      f.depart_date as dep_date,
-      f.depart_time as dep_time, 
-      f.origin_country as dep_country,
-      f.origin as dep_airport_code,
-      f.arrive_date as arr_date, 
-      f.arrive_time as arr_time,
-      f.destination_country as arr_country,
-      f.destination as arr_airport_code,
+      TO_CHAR(f.depart_date, 'DD/MM/YYYY') AS dep_date,
+      TO_CHAR(f.depart_time, 'HH24:MI') AS dep_time, 
+      f.origin_country AS dep_country,
+      f.origin AS dep_airport_code,
+      TO_CHAR(f.arrive_date, 'DD/MM/YYYY') AS arr_date, 
+      TO_CHAR(f.arrive_time, 'HH24:MI') AS arr_time,
+      f.destination_country AS arr_country,
+      f.destination AS arr_airport_code,
       f.airline
     FROM flights f
-    where f.trip_id = $1
+    WHERE f.trip_id = $1
     ORDER BY f.flight_id DESC
   `;
   const res = await query(sql, [trip_id]);
   return res.rows;
 }
 
-export async function delete_flight(trip_id: number, flight_id: number) {
+export async function delete_flight(user_id: string, trip_id: number, flight_id: number) {
+  const check_sql = `
+  SELECT *
+  FROM trip_collaborators
+  WHERE trip_id = $1 AND user_id = $2 AND (role = 'Owner' OR role = 'Editor')
+  LIMIT 1
+  `;
+  const check_res = await query(check_sql, [trip_id, user_id]);
+
+  if (check_res.rowCount === 0) {
+    throw new Error("Permission denied: inviter is not owner/editor");
+  }
+
+
   const sql = `DELETE FROM flights WHERE trip_id = $1 AND flight_id = $2`;
   const res = await query(sql, [trip_id, flight_id]);
   return (res.rowCount ?? 0) > 0; // Will return 1 if remove successfully | Else return 0
@@ -39,7 +52,20 @@ export type FlightInsert = {
   airl_name:string,
 };
 
-export async function post_flight(input: FlightInsert, trip_id: number) {
+export async function post_flight(user_id: string, input: FlightInsert, trip_id: number) {
+  const check_sql = `
+  SELECT *
+  FROM trip_collaborators
+  WHERE trip_id = $1 AND user_id = $2 AND (role = 'Owner' OR role = 'Editor')
+  LIMIT 1
+  `;
+  const check_res = await query(check_sql, [trip_id, user_id]);
+
+  if (check_res.rowCount === 0) {
+    const role = "viewer";
+    throw new Error("Permission denied: inviter is not owner/editor (role: ${role})");
+  }
+
   const sql = `
     INSERT INTO flights (
       trip_id, 
@@ -53,7 +79,7 @@ export async function post_flight(input: FlightInsert, trip_id: number) {
       destination,
       airline
     )
-    VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10)
+    VALUES ($1,TO_DATE($2, 'DD/MM/YYYY'),$3,$4,$5,TO_DATE($6, 'DD/MM/YYYY'),$7,$8,$9,$10)
     RETURNING *;
   `;
 
@@ -74,15 +100,27 @@ export async function post_flight(input: FlightInsert, trip_id: number) {
   return rows[0];
 }
 
-export async function put_flight(input: FlightInsert, trip_id: number, flight_id: number) {
+export async function put_flight(user_id: string, input: FlightInsert, trip_id: number, flight_id: number) {
+  const check_sql = `
+  SELECT *
+  FROM trip_collaborators
+  WHERE trip_id = $1 AND user_id = $2 AND (role = 'Owner' OR role = 'Editor')
+  LIMIT 1
+  `;
+  const check_res = await query(check_sql, [trip_id, user_id]);
+
+  if (check_res.rowCount === 0) {
+    throw new Error("Permission denied: inviter is not owner/editor");
+  }
+  
   const sql = `
-    UPDATE flights 
+    UPDATE flights
     SET
-      depart_date = $1, 
+      depart_date = TO_DATE($1, 'DD/MM/YYYY'), 
       depart_time = $2,
       origin_country = $3,
       origin = $4,
-      arrive_date = $5,
+      arrive_date = TO_DATE($5, 'DD/MM/YYYY'),
       arrive_time = $6,
       destination_country = $7,
       destination = $8,
